hammer/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
Jinja2>=3.0,<4.0
psycopg2-binary>=2.9,<3.0

